Add Not Invented Here logo to supporters section.
I created a sub section of the sponsors area for orgs and people who support the
conference with things like lanyards, free stuff and things of that nature. The
author of the comic strip Not Invented Here sent us a bunch of comics to give
out at the conference (so cool), so that is the first supporter added.