Update site URLs to use {{site.baseurl}}
This makes it so that all URLs to site assets and pages make use of
{{site.baseurl}} so that the assets can be served from the new domain w/
the site as a directory of burlingtonruby.github.io. This change was
made because the burlingtonrubyconference.com domain is expiring, and I
want the site to still be accessible in the future. Before this change
the site was totally broken because the assets wouldn't load.